Build a preliminary search filter system.
Underneath the search bar there are several clickable
"buttons". A js on_click event will add the enabled
filters to the $_POST data. RegEx pattern matching is
used to match the names of the buttons with the
column names used in the database.

If any filters are present, they are passed to the
search query, if they are absent the default filters
are used (which means they are all enabled, i.e. search
everything)